prevent ACK<->RST ping-pong
this relates to the fix for #14.
I forgot to change the sequence number, the ACK will have.
When the ACK is sent, the seq. no needs to be fitting to the first RST the peer sent, so that the nes RST will then hopefully fit THAT sequence no. exactly.
Before this, hte result was a RST<->ACK ping-pong because the peer kept re-sending and the seq no. kept not matching endlessly.